fix: All changes Pushed

+++ src/controllers/employee.controller.js
@@ -1,10 +1,12 @@
 const Employee = require('../models/Employee');
+const SalesOrder = require('../models/SalesOrder');
 const Joi = require('joi');
 const { serverLogger } = require('../utils/loggerWinston');
 const { request } = require('express');
 
 // Joi schema for validation
 const employeeSchema = Joi.object({
+    id: Joi.string().optional(),
     code: Joi.string().required(),
     name: Joi.string().required(),
     mobileNumber: Joi.string().regex(/^\d{10}$/).required().messages({
@@ -12,13 +14,12 @@ const employeeSchema = Joi.object({
     }),
     dateOfBirth: Joi.date().required(),
     address: Joi.object({
-        line1: Joi.string(),
-        line2: Joi.string(),
-        city: Joi.string(),
-        state: Joi.string(),
-        pincode: Joi.string()
-    }),
-    isDeleted: Joi.boolean().default(false)
+        line1: Joi.string().allow(''),
+        line2: Joi.string().allow(''),
+        city: Joi.string().allow(''),
+        state: Joi.string().allow(''),
+        pincode: Joi.alternatives().try(Joi.string().allow(''), Joi.number())
+    })
 });
 
 // Create an employee
@@ -34,9 +35,15 @@ const createEmployee = async (req, res) => {
         if (existingEmployee) {
             return res.status(409).json({ error: 'Employee already exists!.' });
         };
+
+        // Add Explicitly userId and mobileNumber
         body.userId = req.user._id;
         body.mobileNumber = `+91${body.mobileNumber}`;
+        body.dateOfBirth = new Date(body.dateOfBirth);
+
+
         const employee = await Employee.create(body);
+
         res.status(201).json({ msg: 'Employee created successfully', data: employee });
     } catch (error) {
         console.log(error);
@@ -48,8 +55,65 @@ const createEmployee = async (req, res) => {
 // Get all employees
 const getAllEmployees = async (req, res) => {
     try {
-        const employees = await Employee.find({ isDeleted: false, userId: req.user._id });
-        res.status(200).json({ msg: 'All employees', data: employees });
+        const employeeDataWithSoldQuantities = await Employee.aggregate([
+            {
+                $match: { isDeleted: false, userId: req.user._id }
+            },
+            {
+                $lookup: {
+                    from: "salesorders",
+                    let: { employeeId: "$_id" },
+                    pipeline: [
+                        {
+                            $match: {
+                                $expr: { $eq: ["$salesManId", "$$employeeId"] },
+                                isDeleted: false
+                            }
+                        },
+                        {
+                            $unwind: "$products"
+                        },
+                        {
+                            $group: {
+                                _id: null,
+                                totalSold: { $sum: "$products.quantity" },
+                                totalAmount: { $sum: "$totalAmount" }
+                            }
+                        },
+                        {
+                            $project: {
+                                _id: 0,
+                                totalSold: 1,
+                                totalAmount: 1
+                            }
+                        }
+                    ],
+                    as: "salesData"
+                }
+            },
+            {
+                $lookup: {
+                    from: "employees",
+                    localField: "_id",
+                    foreignField: "_id",
+                    as: "employeeObj"
+                }
+            },
+            {
+                $project: {
+                    _id: 1,
+                    name: 1,
+                    code: 1,
+                    mobileNumber: 1,
+                    soldQuantities: { $ifNull: [{ $arrayElemAt: ["$salesData.totalSold", 0] }, 0] },
+                    totalSales: { $ifNull: [{ $arrayElemAt: ["$salesData.totalAmount", 0] }, 0] },
+                    address: { $arrayElemAt: ["$employeeObj.address", 0] },
+                    dateOfBirth: { $arrayElemAt: ["$employeeObj.dateOfBirth", 0] }
+                }
+            }
+        ]);
+
+        res.status(200).json({ msg: 'All employees', data: employeeDataWithSoldQuantities });
     } catch (error) {
         serverLogger("error", { error: error.stack || error.toString() });
         return res.status(500).json({ error: 'Internal Server Error' });
@@ -76,7 +140,11 @@ const updateEmployee = async (req, res) => {
         const { error } = employeeSchema.validate(req.body);
         if (error) {
             return res.status(400).json({ error: error.details[0].message });
-        }
+        };
+
+        const updatedData = req.body;
+        updatedData.mobileNumber = `+91${req.body.mobileNumber}`;
+        updatedData.dateOfBirth = new Date(req.body.dateOfBirth);
 
         const employee = await Employee.findByIdAndUpdate({ _id: req.body.id, isDeleted: false, userId: req.user._id }, req.body, { new: true });
         if (!employee) {
@@ -92,7 +160,7 @@ const updateEmployee = async (req, res) => {
 // Delete an employee
 const deleteEmployee = async (req, res) => {
     try {
-        const employee = await Employee.findByIdAndDelete({ _id: req.body.id, isDeleted: false });
+        const employee = await Employee.findByIdAndDelete({ _id: req.body.id, isDeleted: false, userId: req.user._id });
         if (!employee) {
             return res.status(404).json({ error: 'Employee not found' });
         }
